Revert "Regenerate ART test files (2022-12-05)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e94e86e1-6963-474a-9d0b-319c0791a8e5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ida37cfde24a5c4462ceeba400f0a65460d6dfe18