Push entire test dir in standalone gtests
Simplify standalone gtests xml config.
This also allows pushing multilib test libraries, because FilePusher
can't handle 32/64 suffix if it's not at the end of the file.

For those tests that run generate-boot-image, it is run unconditionally.
I.e. if a boot image already exists after `atest -d`, another `atest`
run of the same test will generate boot image again.

Another point is that generate-boot-image is called for each ABI
directory copied by FilePusher, which happens with `atest -a -d`. In
this case generate-boot-images will be run twice for one of the arches,
because its ABI dir will not be cleaned up.

Bug: 261703065
Test: atest -a art_standalone_(...)_tests
Test: atest -a -d art_standalone_(...)_tests
Change-Id: I63c416b2443f08e317535c1ef517f0575bf624e4