Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b7c540e1-858e-4541-9f0b-3df8fb8f249e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If4b440760abda60f996dd6ba631dca77800e4f9c