Move adding extra goto blocks to InlineInto
There are some cases in which we need to add extra goto blocks
when inlining. We can do that in InlineInto instead of doing
it in the builder which helps decoupling those two stages as
well as simplifying surrounding code.

We also have the benefit of adding the extra goto blocks only
when necessary.

Bug: 227283224
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Change-Id: Ibe21623c94c798f7cea60ff892064e63a38a787a