Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a699c8a2-7399-4986-9f11-fb78ff4ee7e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieaa42b3dda76fa043f365eee0a33dd2cefd88739