Revert "Don't use nterp when instrumentation is active"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5684dea9-2586-4402-bb66-efb73152d95c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iea0b60735cdde80df9d0f57c8bedc4ec51880902