Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0c582f62-f7cb-41c2-950e-5aee388b9da5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I46905985c23b7d2edbd7f8b362571d0a8b319a8b