Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a19e73ea-f46e-41dd-abeb-b8735b875873). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic7d577d23af457f2629ff914e475cd69aa9df0ad