Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/56ea0a2c-c242-4cb0-9589-f791359c7f9d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id925ffc923910740ecd598acd6771860308d502f