Revert "CC GC: Clean up class loaders in reclaim phase."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ab95f36-d69f-4fd1-b6a4-eebeb3fffa62). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifa5096bc0e98e4bbf184e35b5af01b7dec0ad5ef