Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8f4a44fc-84e9-4d9d-882c-e5c3c4d21a9a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4d7cc42560897692856390d92bfd2472732f53f3