Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d0c14d56-fc59-4a6f-a794-155e525b2503). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3dcb0a288e795f1e0c409c4edcb2e832e5c8e278