Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c771bbc-78cb-4fde-aaee-e1e00a87e8c3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I970ddd2a5a6bf449171bdd731d1198d19d8554ec