Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8c5d7b98-19c9-4a0b-80d2-757b99b53cd9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3996aac47e96129ee7f26ca0c03e874dc7d6ad9b