Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/62099b25-5768-4323-b131-648b957c7ddc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I125fe2e71235ffcbe5d8b186b4be9465d95a069c