Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43116086-e7fe-49c5-b89c-755945640b81). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5083ef6cd6f677d580855fa0d5d9a70125ebbd08