Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e0b98d3-90f1-40e7-b833-127dcfc79f6b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I997d6080fd27066e0ae828679b34ea75850bced1