Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83afeb8f-3c53-41d5-b583-8bbe191ca856). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I823358597e1b5a05f0c8ab3ab39dd14e1f2b51da