Update graph flags and check consistency
Check that the flags are up to date in graph checker. Mainly a
correctness CL but it brings slight code size reduction (e.g.
not needing vreg info if HasMonitorOperations is false),
slight compile time speed (e.g. early break in LoopOptimization
if HasLoops is false).

Update loop_optimization_test to stop using `LocalRun` directly
as it meant that it was breaking assumptions (i.e. top_loop_ was
nullptr when it was expected to have a value).

TODO: maybe add the rest of the flags e.g. HasSIMD,
HasBoundsChecks, HasAlwaysThrowingInvokes

Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Change-Id: I29765b3be46d4bd7c91ea9c80f7565a3c88fae2e