[DO NOT SUBMIT] Write-protect sigchain handlers.
A theoretical attack vector against ASYNC and ASYMM MTE is to have a
set of primitives that can overwrite a fuction pointer that's used in a
signal handler. The SIGSEGV in ASYNC and ASYMM (synchronous reads,
asynchronous writes) MTE modes is delivered on the next syscall
boundary, and the actual write to memory succeeds, meaning that the
corrupted function pointer is used.

To protect agains this, wrap the sigchain in a WriteProtected<> class.
This class puts the members in its set of mmap-d pages, which are
usually non-writeable. The chain is only mprotected() to be writeable
temporarily when WriteProtected<>::Mutate is called.

This should help mitigate the theoretical problem.

Test: Boot Android
Bug: N/A
Change-Id: Ia15e8a7b6fdfed0ae532be2734aacd51cdc20576