Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b24dee1f-ec99-4757-8715-500c87d57192). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iedcd814b1693fcee2a33972f3c5c50e7f7e02385