Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d1461cc-7b4d-419e-95c1-b22123f007b4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9b8c45e46d351157eb4144c14f7cbc67bef57084