Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83bc3468-b1fa-4041-b35b-779a5c0e7450). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I21c5b5b91caf23664a2e23a4d1bbcf917c26af51