Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2ba4bc02-5a91-4b54-a365-fda2d3db9501). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I98f5f569ee05ecfac87a8845ae9b0eebb5afae13