Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c3212e3-b2c2-4eeb-be88-59600e48a1e7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If1ab752e114715dd006e76c1b8617caf1532bc75