Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36592af7-9efc-46d8-81b3-3560c7c31e93). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I724e7e325fa7b8fa346f9a1feefae3d6a25dd225