Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9b904657-d359-4996-93ac-0f62ca29300b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I72d7578ddcff309fba696f944f0bdce5ac0b4391