Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21199f8d-91f7-4bea-b0cd-5acc38a28e53). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iac663ffc4f98992d338e27aff198ec50199b5077