Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d07a208c-460b-41ee-9461-fc89a4b47814). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifdf8ed3adaedcadfb52403282c9aed9e8c83cf0c