Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/65b6d041-5563-40e6-a88f-26c779539e0b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic481e80c7cfd7f43a47cb0b9fdbaf0a6d36ac4bf