Take super-classes into consideration for from-space freeing
For cases where the reference bitmap in the class can't hold all the
references, we iterate over super classes while visiting references.
In such cases we need to ensure that super classes, which may get
allocated after the derived class in address order, are also not freed
before the lowest address order object is compacted.

Test: module install and enable userfaultfd GC
Bug: 160737021
Change-Id: I4ca581d5fa66f6ded983ff627d56adb6e8db6177