Remove duplicates in ART apex
bootclasspath_fragment causes the dependent libs to be in the apex.
Having those libs in java_libs as well is confusing because
bootclasspath_fragment might do something with those deps but it's not
clear which artifacts are used in the apex.

Now, it's clear that those dependencies are from bootclasspath_fragment,
not from java_libs.

Similarly, test_broken_com.android.art is now explicit about which
version of libart is used. Previously, it was not clear that
libart-broken is used. (Note that, even if libart is not listed in
native_shared_libs, the build system traces the dependency graph to it
because it's used by libart-compiler.)

Bug: 191770320
Test: m com.android.art
Test: m test_broken_com.android.art
  (it should have libart-broken)
Change-Id: I17df5fc25c1d518e38155509a8b111545ec4305a