Revert "Skip `VMStack.getThreadStackTrace()` in stack traces."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd814d88-67ef-4793-a343-5c2c353ae90a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I60e7c7c8f01f76971388ff5880ed5d784aa1f930