Revert "Skip `VMStack.getThreadStackTrace()` in stack traces."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/795a3db8-7d8b-42ea-8983-2d08a0d7fa3d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4380c5be9cfc0c8827e463085c131c36cf6d9dc7