Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b52e5d3-6df9-4c5d-baff-f8c44dc5fd35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifcbd16e3fb9eb47319b0c1ff7c7351025bf1e07c