Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1567fece-d938-4ae4-bd93-ce4435adeca9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f867f92760e6c22c71abcd1d0a589d87a05b069