Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7734994c-b379-42e3-bc3c-6421ed58bc1e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I77fcfaa023f42ecdc766b9b4329e2a85646385eb