Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/12521240-b327-461e-a335-bd8bc1ae1450). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e8d29f08434b7cf7bad5821c1379bb16a88d6db