Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/beb79034-efca-4526-ade3-06941748f8d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3fa65537a3151e3953790a19fbb94a1eb93cfa4e