Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/284d4f5c-c196-4f3a-9b2d-df2421f87468). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2e5d0864a2413cae1508ba6f215fe1880fe9f60e