Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/19c692ca-e654-4215-a0b9-a882fedd17a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I39841449f4b70f6c44967332964cc26fbf4dbfc8