Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d578fe50-f9c2-47cc-a051-8cb134e53dae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I617bfff6f5a99ba0c2f0fd3bccd3e229b5f22c91