Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/65502955-6daf-4e08-95ed-49cd30d06374). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7d1a8582608df0eb65d80eb1e75f6459f9072d67