Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b148a230-43cd-48b6-b8ea-4ba3147a69b2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibb4c75ea32a6b4e82291a0b30abcdfa1547c57d8