Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ada56ec3-2a0b-4a11-b783-3beb827d935d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I24f63a3638c172759581fe40140a6c56d97acfee