Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/378b6c4a-0b78-453e-baae-ec72c5e3b04d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I77be11bdd5366b1e0f6488c2f70407aa91e2953c