Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0dd7f89d-23fe-41b8-ae8e-649e3b3f8a0f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7cd96b3f2d9a85c2e9890675b08657d42531689d