Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8d8c2fb9-4904-4e14-9cf2-d42a4c9733ed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I53264febd170bbc0c0ffa5e1054c7f6ab5cc561e