Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9215f2e5-96c8-4204-8777-2c2fd277ffeb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2f5656adbee0fddca2ec946195dea446c4e9bcbe