Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8ee97e18-f7c2-4bb8-ad80-bd20e08422a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I048cbf9fe7ebc489084bdb4a0c390f2303ea954e