Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4dc81819-a94b-480e-af1e-5024af56c4ee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icd072caf3603d15294d67bb53b3c24943ba0d73e