Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/57e28dc0-64a5-41df-a702-9a5cab4348cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3c147db15f07642f5c065660714cd80ae4b360df