Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/facfed7f-65c4-4284-9f8e-cb009f07830f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If22fa10ae736cf94154b6eb15c3065e6c2abfc56