Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3a2c4b43-9cfe-4091-9c62-5bacf336d25e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I836ed9b977186146df54519d90f6785290d55699