Revert "Clear mark-bitmap after compaction"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/673ad92b-3be0-4fe7-b832-2ab91fa2cfa8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I970091b6e9283e6beb45e57390726f45db178e4a