Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae679a00-c75a-408c-bd67-9c5fc37c937f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3f0010efd3862d5e8693a818a66a58665c95b817