Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1d246ce6-c9c1-41da-981f-85455b6c23a5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idbcbd9b516556548ec2d34cdf6eb921d6c3fee6e