Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b691bcf2-423d-49ac-b32b-49ebf8dffe87). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I10f343cd62f049bc9d047590d0537ec77175e604