Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/648923d0-ef2c-43fe-af7f-02f888928ea2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic49da5ce10ad8bbdbd3b94ef5d64d3ffe705b066