Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c1aaefbb-23c4-4b09-8e42-77f602947a62). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7fafddecbc52ac9218fe835edff74c1d53351bb7