Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ba061bd8-065b-4872-8176-3f79b985c295). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifcd8f3bf1151db53b964481949e0ff9b06ed9944