Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dbfdf388-22e4-4bbb-b5d1-a7fabac81443). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibfe39d23db8a3e120f107e1fc24b896b06545e3e