Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cc4bc175-1da3-417b-b83c-241e9b4dbdea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2743aa49c4fb9d2d406c57143657b6f0bb42b488