Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4711e3d9-46a9-4db3-aeb5-4ae853f6727f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie39a67e0b7cfe8d2488e606c8642c289faac0e6b