Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/729e30e1-d580-4aa4-937d-fe42a66545ae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie82a16f3e2109741d4fc9929f3f39a61e1cccdcc