Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a34284ea-9666-48b6-995b-6747bff8312d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I742344216861cfea2dedbd27c38845cd3eb8f800