Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/059ad29c-ef19-4610-b2cf-1d627d01833c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie4e5dca045694f7681fd536e1534d166e20e8e11