Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ebe2c0d-fb05-43fe-a7ff-ce6bcbe4e77d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I97911116a980d4c4d224526ae747cfddd3a4ed71