Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1a56e9f-2de9-41e4-ba1d-0295e054ebd9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie9979dade63326eafccb6434d8f854fdc87fb826