Revert "Clear mark-bitmap after compaction"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/00a513ff-21cf-4692-b522-a81e1a2d5d32). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I766e1c024e8e8e1073b8bec5b715f78ccbddc902