Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb741bcd-4dfd-4c98-88fe-8f7865e4ec4b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia1e9a5f6c7532bd7fa5b5da11c4dbb2c52107bf3