Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/38226fee-0cb4-4f2f-80a2-577ebb919606). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id2b1860844801602d12c1e5f153d599987baada2