Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7a565b83-ea17-4ab8-9c3e-381d7e208a10). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I05bce219fb3c417ab273f808164a64f3c2442d2c