Revert "Replace blkid with blkid_static"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b0875994-12cd-4f2c-8061-18dd5e8b1dc3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9085729bb0258e9de2a9a55d577690d5ed950df8