Cleanup handling of frame pop debug feature
When a frame pop is requested we need to ignore any existing exception
and pop the existing frame. This is all handled in interpreter and if we
see a frame pop request we simply deoptimize and let interpreter handle
it. Though before deoptimizing we used to clear the exception which
required additional checks and additional stack walks. Interpreter
already handles supressing and clearing exceptions on frame pop requests
so we can cleanup the code in QuickExceptionHandler and
QuickToInterpreterBridge to not special case frame pop requests by
clearing exceptions.

Bug: 206029744
Test: art/test.py

Change-Id: I49848145ba811387eb09fd7ae3a1020a4b9bd202