Disable test 004-SignalTest in switch interpreter mode.
This test relies on sigaction calls being intercepted and handled by
ART. If ART fault manager is not installed (which may be the case in
switch interpreter mode), then sigaction calls go to the kernel and the
test behaves unexpectedly (signals SIGKILL and SIGSTOP in the mask are
unblocked, while the test expects them to be blocked).

Prevously the test didn't fail on the switch interpreter, as all
currently supported architectures happen to install ART fault manager.

Bug: none
Test: art/test.py --target -r --64 --interpreter 004-SignalTest
      # observe that the test is skipped
Test: art/test.py --target -r --64 --optimizing 004-SignalTest
      # observe that the test runs successfully
Change-Id: Ie8c1cf3f5add15ccc1225c2147d13b5021ebb25c