Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/62a738a6-bb0f-415e-b7f5-fe2210ea3000). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5f4b8c43cd45b01bf982a55a86d1783ee43ae395