Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b344bd53-ecd8-4f73-b16b-4a38b1a9bc0d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I435a28d266a658d556b43478a25960cbc86e1e31