Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4b5f4aae-8ee8-498a-b89d-da980167e322). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id52b8fb02cba5acd82f9498caef1ccc24a03fdfe