Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1274ddb-7941-4bb1-b727-06df2e8a1ca8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia8cce43a04e5b2481b5064b5df6cf9cc97efe2be