Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81fd85e1-e4a3-49d7-af2e-667381e594d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6d64b81dab77bc47f1fd4a30538f2e5a2b8c7fa7