Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e3069a1b-8f28-44e3-b6c6-495577e7c222). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id72e8e77ff9e7e8671f044d8b9d60d17ebee22f2