Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20c2e729-3146-45e2-a0d7-192eab36595e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic2e925e073f98f34f87eb6180e8d13a5514acdfd