Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8fc6d4e7-6dce-4ecc-9a56-62c052ce6016). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7dcdbd68ed8ecde48c76a2a5408f4400b257f3fc