Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/772c3553-b3ce-4e70-aad7-e9bbcc861922). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I711432d1a5ebea8e7bfef51ca7d56850d75f5d4a