Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/90541c78-e347-45eb-89b1-b879871d2957). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic649a7e52b816f0c079e36521fe0fb49cc5cb963