Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca59e1b4-dccd-4ce4-8208-dd1e7828b9a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I265a1d0c04e54b348921c42e391d66b353bcb4b5