Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a72c0543-1070-4a3c-9cad-99f10ebd4680). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8dc27f3d4baf9f008b649b9e88f84a41fdbea435