Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a0594086-02ae-43f9-a85d-c7001e3ffdb7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8bf931a91ceee8746d28caf9b90d6d287e2cde68