Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b907e833-9dd8-4037-a384-a72c32732159). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I38fc7c83a58249c504b1eb0a76e23069234cf200