Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0a1f876a-32c7-40f7-a1e9-76317b94ee15). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie5152bf8fb6bffd998b598b76bab58de6e5f365b