Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/61bed13a-f5d9-4b05-977f-ec6b8a704f29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic31bc1b0f1eaac78f13a0c8ade82bc3803a4c843