Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2cbf4ffc-bc01-477e-a7e8-ebd7894dab5d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1ebeabf1df25ffa8d9f977875aad13b9212d2922