Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a85dfed4-ff85-4eb8-9ea6-f6af4a6fe2c7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I25f923e650a91798533b955732004e7d03ea3dbc