Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb64cabf-91dc-499c-b347-adf76bd60bf5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5137622d4d9d9ecec9db07f45c8c33c450a1d09c