Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0b35cd5b-a4fb-4b70-bad7-f087ffdbd8d2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8fc393653d8b9b90151e2c6256957d6651c0d563