Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c6e56a25-44ec-4c7d-83e2-3f43a490a27c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7179d6b25f4872b961a8e61b6ab66d597e147ab9