Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1acca68e-47ef-4098-a42a-2756064df3e7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe13eddecc1da4230598ac1fd5cef4013e94bcb1