Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7fffd60a-0963-45c0-b063-d88d294940a1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibc7ccfc418fbca2051a69e95b1d8be9875b14e5b