Revert "Write classes in runtime-generated app image."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d2834fb-ff65-4ae7-a31f-98890341641a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1041f2b24baff90ce060cdd1799c989b4c7defd2